organize http module + add doc blocks to HTTP body types
