--- Sources/WebSocket/Exports.swift
@@ -1,3 +1,0 @@
-@_exported import HTTP
-@_exported import NIO
-@_exported import NIOWebSocket
